Haribo jellies (last scene) detection enhenced
There was a small mistake in jellies detection in the last scene. If
program couldn't find exactly 4 markers, it proceeded to the next
picture of the scene. The problem occured if the program detected
different number of markers on the last picture then the loop was
continued without saving the results.